Remove type hack
When I was working on the RPC service thest first time around I
encountered a compiler bug and the solution was to implement a trait on
a tuple including a type, not a type itself. Sadly I haven't linked to
the description of the source of the bug, but it looks like it's no
longer a bug as of Rust 1.76, thus I'm removing the hack.