Allow concurrent requests to RPC services
The initial implementation didn't allow requests to services to be
handled concurrently. For a proof of concept or a single scenario run
it's totally fine, but in the future it may make a lot of sense to
deploy Crows as a long running service and leave it running with workers
connected and send requests to Crows from a CI run and then multiple
runs may be sending multiple requests concurrently.