Allow to construct a service after creating a client
Before this commit service had to be constructed before the client was
even created. Thus the client couldn't be also used (for example to
request information from the other side of the connection) before
constructing the service. This commit changes the signature of the
connection functions to accept create_service_callback and allow to
construct the service after the client is available.

Another advantage, on top of being able to request data before creating
the service, is that now it's possible to add the client to the service
when it's needed and not necessarily always inject it