Send updates as channel messages instead of polling
The whole point of allowing to send messages both ways was to be able to
call functions on both sides of the RPC connection. It was at first
mostly used in the woker <-> coordinator collection as the workers
connect to coordinator and then it's mostly coordintor that sends
messages, but when I was writing run updates I used polling as I was in
a hurry and handling clients in services was not straightforward. After
changing the services code it's now trivial to put client in a service
and thus pushing data from workers all the way to the client is also
relatively easy.