Fix flushing output messages
Messages that come from the worker are not necessarily lines, ie. they
may be just a few characters out of a full line. If we were printing
them as they come, we could just pass them to the output, but we want to
prefix each line with [ERROR] or [INFO] and the name of the worker.
That's why we need to buffer the data and extract lines or flush after a
certain amount of time has passed